pyproject.toml
[tool.poetry]
name = "final-project"
version = "0.1.0"
description = ""
authors = ["Oz"]
readme = "README.md"
packages = [{ include = "app" },
            {include = "utils", from = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
folium = "^0.18.0"
geopy = "^2.4.1"
numpy = "^2.1.3"
plotly = "^5.24.1"
scikit-learn = "^1.5.2"
seaborn = "^0.13.2"
matplotlib = "^3.9.2"
imbalanced-learn = "^0.12.4"
pyaml = "^24.9.0"
mlflow = "^2.18.0"
torch = "^2.5.1"
xgboost = "^2.1.2"
pyyaml = "^6.0.2"
streamlit = "^1.40.1"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
